refactor: created custom api routing
