+++ apps/web/app/components/signIn/SignInWrapper.tsx
@@ -29,5 +29,7 @@ export const SignInWrapper = ({ children }: { children: React.ReactNode }) => {
 			</AlertDialog>
 		);
 	}
-	return <a href={process.env.CF_PAGES_URL + "/auth/google"}>{children}</a>;
+	return (
+		<a href={process.env.CF_PAGES_URL + "/api/auth/google"}>{children}</a>
+	);
 };
