+++ apps/web/app/routes/_app/$handle/index.tsx
@@ -88,7 +88,7 @@ const SignOutButton = () => {
 		<Button
 			variant="outline"
 			onClick={() => {
-				fetch("/auth/signout").then(() => {
+				fetch("/api/auth/signout").then(() => {
 					navigate({
 						to: "/",
 					});
